routes: migrate cost domain (5 routes)
Move to src/server/routes/cost.ts:
- /api/sessions/:id/cost (GET)
- /api/sessions/:id/cost/breakdown (GET)
- /api/goals/:id/cost (GET)
- /api/goals/:id/cost/breakdown (GET)
- /api/tasks/:id/cost (GET)

Net: server.ts -155 lines (5819 → 5664). No behaviour change.
