+++ src/server/routes/dispatcher.ts
@@ -28,6 +28,7 @@ import { workflowsRoutes } from "./workflows.js";
 import { staffRoutes } from "./staff.js";
 import { mcpRoutes } from "./mcp.js";
 import { maintenanceRoutes } from "./maintenance.js";
+import { costRoutes } from "./cost.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -47,6 +48,7 @@ const allRoutes: ReadonlyArray<Route> = [
 	...staffRoutes,
 	...mcpRoutes,
 	...maintenanceRoutes,
+	...costRoutes,
 ];
 
 /**
